--- a/tests/Feature/Http/Api/ApiHostgroupsTest.php
+++ b/tests/Feature/Http/Api/ApiHostgroupsTest.php
@@ -2,20 +2,23 @@
 
 namespace Tests\Feature\Http\Api;
 
-use App\Enums\Roles;
 use App\Models\Host;
 use App\Models\Hostgroup;
 use App\Models\User;
 use Tests\ApiTestCase;
+use Tests\Traits\InteractsWithPermissions;
 
 class ApiHostgroupsTest extends ApiTestCase
 {
+    use InteractsWithPermissions;
+
     private User $user;
 
     public function setUp(): void
     {
         parent::setUp();
-        $this->disableAuthorization();
+
+        $this->disablePermissionsCheck();
         $this->user = User::factory()->create();
     }
 
@@ -47,7 +50,7 @@ class ApiHostgroupsTest extends ApiTestCase
             ],
             'relationships' => [
                 'hosts' => [
-                    'data' => $hosts->map(fn (Host $host) => [
+                    'data' => $hosts->map(fn(Host $host) => [
                         'type' => 'hosts',
                         'id' => (string) $host->getRouteKey(),
                     ])->all(),
@@ -165,7 +168,7 @@ class ApiHostgroupsTest extends ApiTestCase
             ],
             'relationships' => [
                 'hosts' => [
-                    'data' => $newHosts->map(fn (Host $host) => [
+                    'data' => $newHosts->map(fn(Host $host) => [
                         'type' => 'hosts',
                         'id' => (string) $host->getRouteKey(),
                     ])->all(),