--- a/tests/TestCase.php
+++ b/tests/TestCase.php
@@ -6,20 +6,12 @@ use Database\Seeders\PermissionsTableSeeder;
 use Database\Seeders\RolesTableSeeder;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Testing\TestResponse;
 
 abstract class TestCase extends BaseTestCase
 {
     use CreatesApplication;
-    use RefreshDatabase;
-
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->seed(PermissionsTableSeeder::class);
-        $this->seed(RolesTableSeeder::class);
-    }
 
     /**
      * Make ajax GET request.