--- a/tests/Traits/InteractsWithPermissions.php
+++ b/tests/Traits/InteractsWithPermissions.php
@@ -49,5 +49,3 @@ trait InteractsWithPermissions
         return $this;
     }
 }
-
-