**What issue type does this pull request address?** (keep at least one, remove the others)  
/kind enhancement

**What is this pull request for? Which issues does it resolve?** (use `resolves #<issue_number>` if possible)  
resolves #166 


**Does this pull request has user-facing changes?** (e.g. config changes, new/modified commands, new/modified flags)  
Yes, It increase testing performance

**Before**:
Tests:  150 passed  Time:   66.59s
Tests:  150 passed  Time:   69.94s

**After**:
Tests:  150 passed  Time:   21.14s
Tests:  150 passed  Time:   21.31s