@pacoorozco 2021-06-18
After all the refactors, this is the final time. 

Tests:  170 passed  Time:   29.61s

And this is the final coverage: